# frontend/packages/api-client/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "Consistent error contract distinguishing validation/business/conflict/infrastructure outcomes.",
        "properties": {
          "detail": {
            "description": "Specific explanation for this occurrence.",
            "title": "Detail",
            "type": "string"
          },
          "title": {
            "description": "Short human-readable summary.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "Stable machine-readable error category, e.g. 'not_found'.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "OrgaRoleResponse": {
        "properties": {
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "displayNameChain": {
            "items": {
              "type": "string"
            },
            "title": "Displaynamechain",
            "type": "array"
          },
          "entityId": {
            "title": "Entityid",
            "type": "string"
          },
          "entityKind": {
            "const": "OrgaRole",
            "default": "OrgaRole",
            "title": "Entitykind",
            "type": "string"
          },
          "properties": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AirlineRoleProperties"
              },
              {
                "$ref": "#/components/schemas/EmptySupplierRoleProperties"
              }
            ],
            "title": "Properties"
          },
          "schemaVersion": {
            "title": "Schemaversion",
            "type": "integer"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "entityId",
          "type",
          "schemaVersion",
          "properties",
          "displayName",
          "displayNameChain"
        ],
        "title": "OrgaRoleResponse",
        "type": "object"
      },
...
      "Page_ProductResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProductResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[ProductResponse]",
        "type": "object"
      },
...
      "ProductCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "entityId": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entityid"
          },
          "parentProductId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parentproductid"
          },
          "product": {
            "discriminator": {
              "mapping": {
                "product/accommodation/room-type": "#/components/schemas/RoomCategoryRequest",
                "product/airline/flight": "#/components/schemas/FlightRequest",
                "product/experience/activity": "#/components/schemas/EmptyProductRequest",
                "product/experience/guided-tour": "#/components/schemas/EmptyProductRequest",
                "product/mobility/coach": "#/components/schemas/EmptyProductRequest",
                "product/mobility/rail": "#/components/schemas/EmptyProductRequest",
                "product/mobility/transfer": "#/components/schemas/EmptyProductRequest",
                "product/mobility/vehicle-rental": "#/components/schemas/EmptyProductRequest",
                "product/protection/travel": "#/components/schemas/EmptyProductRequest",
                "product/water-transport/cruise": "#/components/schemas/EmptyProductRequest",
                "product/water-transport/day-boat": "#/components/schemas/EmptyProductRequest"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/FlightRequest"
              },
              {
                "$ref": "#/components/schemas/RoomCategoryRequest"
              },
              {
                "$ref": "#/components/schemas/EmptyProductRequest"
              }
            ],
            "title": "Product"
          }
        },
        "required": [
          "product"
        ],
        "title": "ProductCreateRequest",
        "type": "object"
      },
      "ProductMutationResponse": {
        "properties": {
          "entityId": {
            "title": "Entityid",
            "type": "string"
          },
          "entityKind": {
            "const": "TouristicProductItem",
            "default": "TouristicProductItem",
            "title": "Entitykind",
            "type": "string"
          },
          "properties": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FlightProperties"
              },
              {
                "$ref": "#/components/schemas/RoomCategoryProperties"
              },
              {
                "$ref": "#/components/schemas/EmptyProductProperties"
              }
            ],
            "title": "Properties"
          },
          "schemaVersion": {
            "title": "Schemaversion",
            "type": "integer"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "entityId",
          "type",
          "schemaVersion",
          "properties"
        ],
        "title": "ProductMutationResponse",
        "type": "object"
      },
      "ProductResponse": {
        "properties": {
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "displayNameChain": {
            "items": {
              "type": "string"
            },
            "title": "Displaynamechain",
            "type": "array"
          },
          "entityId": {
            "title": "Entityid",
            "type": "string"
          },
          "entityKind": {
            "const": "TouristicProductItem",
            "default": "TouristicProductItem",
            "title": "Entitykind",
            "type": "string"
          },
          "properties": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FlightProperties"
              },
              {
                "$ref": "#/components/schemas/RoomCategoryProperties"
              },
              {
                "$ref": "#/components/schemas/EmptyProductProperties"
              }
            ],
            "title": "Properties"
          },
          "schemaVersion": {
            "title": "Schemaversion",
            "type": "integer"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "entityId",
          "type",
          "schemaVersion",
          "properties",
          "displayName",
          "displayNameChain"
        ],
        "title": "ProductResponse",
        "type": "object"
      },
      "ProductUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "product": {
            "discriminator": {
              "mapping": {
                "product/accommodation/room-type": "#/components/schemas/RoomCategoryRequest",
                "product/airline/flight": "#/components/schemas/FlightRequest",
                "product/experience/activity": "#/components/schemas/EmptyProductRequest",
                "product/experience/guided-tour": "#/components/schemas/EmptyProductRequest",
                "product/mobility/coach": "#/components/schemas/EmptyProductRequest",
                "product/mobility/rail": "#/components/schemas/EmptyProductRequest",
                "product/mobility/transfer": "#/components/schemas/EmptyProductRequest",
                "product/mobility/vehicle-rental": "#/components/schemas/EmptyProductRequest",
                "product/protection/travel": "#/components/schemas/EmptyProductRequest",
                "product/water-transport/cruise": "#/components/schemas/EmptyProductRequest",
                "product/water-transport/day-boat": "#/components/schemas/EmptyProductRequest"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/FlightRequest"
              },
              {
                "$ref": "#/components/schemas/RoomCategoryRequest"
              },
              {
                "$ref": "#/components/schemas/EmptyProductRequest"
              }
            ],
            "title": "Product"
          }
        },
        "required": [
          "product"
        ],
        "title": "ProductUpdateRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Shared capability-oriented API reused by Customer and Staff Interaction.",
    "title": "Christopher Columbus Travel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/products": {
      "get": {
        "operationId": "listProducts",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Opaque cursor from a previous page's nextCursor.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from a previous page's nextCursor.",
              "title": "Cursor"
            }
          },
          {
            "description": "Filter by exact terminology type, e.g. product/airline/flight",
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by exact terminology type, e.g. product/airline/flight",
              "title": "Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProductResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "List Products",
        "tags": [
          "products"
        ]
      },
      "post": {
        "operationId": "createProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Create Product",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{product_id}": {
      "delete": {
        "operationId": "deleteProduct",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Delete Product",
        "tags": [
          "products"
        ]
      },
      "get": {
        "operationId": "getProduct",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Get Product",
        "tags": [
          "products"
        ]
      },
      "put": {
        "operationId": "updateProduct",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Update Product",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{product_id}/ancestors": {
      "get": {
        "description": "Root-first CONTAINS parent chain, excluding `product_id` itself; empty when it is already a root.",
        "operationId": "getProductAncestors",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProductResponse"
                  },
                  "title": "Response Getproductancestors",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Get Product Ancestors",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{product_id}/components": {
      "get": {
        "operationId": "getProductComponentTree",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProductComponentResponse"
                  },
                  "title": "Response Getproductcomponenttree",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Get Product Components",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{product_id}/supplier": {
      "get": {
        "operationId": "getProductSupplier",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/OrgaRoleResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Getproductsupplier"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Get Product Supplier",
        "tags": [
          "products"
        ]
      },
      "put": {
        "operationId": "setProductSupplier",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSupplierRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Set Product Supplier",
        "tags": [
          "products"
        ]
      }
    },
...
  }
}